# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Учебный продакшн-подобный сервис учёта клиентов.\n\nКлючевая механика: обычный пользователь не редактирует карточку\nклиента напрямую — он ПРЕДЛАГАЕТ правку, которая применяется\nтолько после одобрения администратором.\n\nДемо-учётки (заводятся миграцией 00004):\nadmin@example.com / admin12345 — роль admin\nuser@example.com / user12345 — роль user",
        "title": "Clients API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/suggestions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Все предложения всех авторов. Доступно только администратору.\nОбычный сценарий: ?status=pending — что ждёт рассмотрения.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Очередь модерации",
                "parameters": [
                    {
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по клиенту (UUID)",
                        "name": "client_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Записей на страницу (1..100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Сколько пропустить",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ListResponse-github_com_vitikevich-landau_clients-api_internal_model_SuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/suggestions/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Атомарно применяет payload к карточке клиента и переводит\nпредложение в статус approved. Только для администратора.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Одобрить предложение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID предложения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Необязательный комментарий",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ApproveSuggestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.SuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Предложение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Уже рассмотрено, клиент удалён или email занят",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/suggestions/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Переводит предложение в статус rejected. Комментарий\nобязателен — автор должен понимать причину отказа.\nТолько для администратора.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Отклонить предложение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID предложения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Причина отказа",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.RejectSuggestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.SuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Не указан комментарий",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Уже рассмотрено",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/clients/{id}/suggestions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Карточка клиента НЕ меняется. Создаётся предложение\nсо статусом pending, которое рассмотрит администратор.\nТело — частичное изменение по правилам JSON Merge Patch:\nзначение = заменить, null = очистить, отсутствие ключа = не трогать.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suggestions"
                ],
                "summary": "Предложить правку клиента",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Предлагаемые изменения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.SuggestionPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.SuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Пустой или невалидный payload",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Клиент не найден",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "У вас уже есть нерассмотренное предложение по этому клиенту",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/suggestions/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Показывает только предложения, созданные текущим пользователем,\nвместе с их статусами и комментариями модератора.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suggestions"
                ],
                "summary": "Мои предложения",
                "parameters": [
                    {
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по клиенту (UUID)",
                        "name": "client_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Записей на страницу (1..100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Сколько пропустить",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ListResponse-github_com_vitikevich-landau_clients-api_internal_model_SuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/suggestions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обычный пользователь видит только свои предложения.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suggestions"
                ],
                "summary": "Одно предложение",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID предложения",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.SuggestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Не найдено либо принадлежит другому пользователю",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_vitikevich-landau_clients-api_internal_model.ApproveSuggestionRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string",
                    "maxLength": 1000
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.ListResponse-github_com_vitikevich-landau_clients-api_internal_model_SuggestionResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.SuggestionResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Pagination"
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.RejectSuggestionRequest": {
            "type": "object",
            "required": [
                "comment"
            ],
            "properties": {
                "comment": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.SuggestionPayload": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Optional-string"
                },
                "email": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Optional-string"
                },
                "full_name": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Optional-string"
                },
                "notes": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Optional-string"
                },
                "phone": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Optional-string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Optional-github_com_vitikevich-landau_clients-api_internal_model_ClientStatus"
                }
            }
        },
        "github_com_vitikevich-landau_clients-api_internal_model.SuggestionResponse": {
            "type": "object",
            "properties": {
                "author_email": {
                    "description": "Поля из связанных таблиц. Заполняются не всегда,\nпоэтому omitempty — чтобы не слать пустые строки без нужды.",
                    "type": "string"
                },
                "author_id": {
                    "type": "string"
                },
                "changed_fields": {
                    "description": "ChangedFields — плоский список имён полей.\nДублирует информацию из payload, но позволяет фронтенду\nнарисовать \"предложено изменить: телефон, заметки\"\nбез разбора структуры payload.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "client_id": {
                    "type": "string"
                },
                "client_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payload": {
                    "description": "Payload сериализуется нашим MarshalJSON — то есть содержит\nтолько те поля, которые реально предложены к изменению.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.SuggestionPayload"
                        }
                    ]
                },
                "review_comment": {
                    "type": "string"
                },
                "reviewed_at": {
                    "type": "string"
                },
                "reviewer_email": {
                    "type": "string"
                },
                "reviewer_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.SuggestionStatus"
                }
            }
        },
        "github_com_vitikevich-landau_clients-api_internal_model.SuggestionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-comments": {
                "SuggestionStatusApproved": "одобрено и применено",
                "SuggestionStatusPending": "ждёт модерации",
                "SuggestionStatusRejected": "отклонено"
            },
            "x-enum-descriptions": [
                "ждёт модерации",
                "одобрено и применено",
                "отклонено"
            ],
            "x-enum-varnames": [
                "SuggestionStatusPending",
                "SuggestionStatusApproved",
                "SuggestionStatusRejected"
            ]
        },
...
        "github_com_vitikevich-landau_clients-api_internal_response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code — машиночитаемый код: not_found, validation_error, ...\nКлиент принимает решения ПО КОДУ, а не по тексту:\nтекст можно переписать или перевести, код — контракт.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_apperr.Code"
                        }
                    ]
                },
                "details": {
                    "description": "Details — пояснения по конкретным полям при ошибке валидации.\nomitempty: в остальных случаях ключа в JSON просто не будет.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Message — человекочитаемое описание. Безопасное: без имён таблиц,\nSQL и стектрейсов.",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID — чтобы пользователь мог назвать его в поддержке,\nа мы по нему нашли весь запрос в логах.",
                    "type": "string"
                }
            }
        },
        "github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorBody"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите: Bearer \u003cваш JWT\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}